+++ CommonLibSF/include/RE/T/TESContainer.h
@@ -1,10 +1,22 @@
 #pragma once
 
+#include "RE/B/BSTArray.h"
 #include "RE/B/BaseFormComponent.h"
 
 namespace RE
 {
-	class ContainerObject;
+	class TESBoundObject;
+	struct ContainerItemExtra;
+
+	struct ContainerObject
+	{
+	public:
+		// members
+		std::int32_t        count;      // 00
+		TESBoundObject*     obj;        // 08
+		ContainerItemExtra* itemExtra;  // 10 - COED
+	};
+	static_assert(sizeof(ContainerObject) == 0x18);
 
 	class TESContainer : public BaseFormComponent
 	{
@@ -18,8 +30,7 @@ namespace RE
 		void                 InitializeDataComponent() override;     // 02 - { return; }
 
 		// members
-		ContainerObject** containerObjects;     // 08
-		std::uint32_t     numContainerObjects;  // 10
+		BSTArray<ContainerObject> containerObjects;  // 08
 	};
 	static_assert(sizeof(TESContainer) == 0x18);
 }
